Refactor InjectIntoHtml and related components
- Introduced ImageExtensions class for image handling.
- Added JavaScriptConstants for centralized JS snippets.
- Created StyleSheetConstants for managing CSS styles.
- Improved code modularity and readability by utilizing new classes.
- Streamlined regex patterns and methods for better performance.
- Removed redundant methods and enhanced overall structure.
